circular cash bugfix # Conflicts: #	.idea/workspace.xml #	out/production/ATSP/ATSP.class #	out/production/ATSP/AtspGreedy.class #	out/production/ATSP/AtspRandom.class #	out/production/ATSP/Main.class

+++ src/TSP.java
@@ -227,7 +227,7 @@ public abstract class TSP {
             solutions.add(new Solution(currentSolution));
         }
         for (int i = 0; i < times; i++) {
-            Long similarity= Long.valueOf(0);
+            Double similarity= Double.valueOf(0);
             for (int j = 0; j < times; j++) {
                 if(i!=j){
                     similarity+=solutions.get(i).calculateSimilarity(solutions.get(j));
